backend/requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
fastapi
uvicorn
google-cloud-aiplatform 
google-cloud-texttospeech
langchain-google-vertexai